+++ lib/main.c
@@ -28,7 +28,7 @@ int flosslogic_hw_init(int hw, struct flosslogic_context *ctx)
 }
 
 uint8_t *flosslogic_hw_get_samples(int hw, struct flosslogic_context *ctx,
-				   uint64_t numsamples, uint8_t samplerate)
+				   uint64_t numsamples, uint64_t samplerate)
 {
 	return flosslogic_logic_analyzers[hw].get_samples(ctx, numsamples,
 							  samplerate);
